Remove radius from Applicant Search Form backend
The Applicant Search Form must not have a setting for the radius value,
since it is included in the address searchcriteria already. This
prevents a range (website) vs. range (search criteria) conflict.

Since both settings for the order would be equal after removing the
radius, only one option would have been left, which is why the `order`
option was removed as well for now.

closes onOffice ticket #1422931